+++ fontbe/src/test_util.rs
@@ -9,7 +9,7 @@ use fontdrasil::{
 use write_fonts::types::Tag;
 
 #[cfg(test)]
-pub(crate) fn axis(min: f32, default: f32, max: f32) -> Axis {
+pub(crate) fn axis(min: f64, default: f64, max: f64) -> Axis {
     let mut mappings = Vec::new();
     if min < default {
         mappings.push((UserCoord::new(min), DesignCoord::new(min / 10.0)));
